+++ shapes/cylinder.h
@@ -6,7 +6,7 @@
 class Cylinder : public OpenGLShape
 {
 public:
-    Cylinder(int p1, int p2, int p3);
+    Cylinder(int p1, int p2, float p3);
 
     virtual void reCalculateVertices();
 };
